clean up list
