+++ pyo3-object_store/src/azure/store.rs
@@ -2,7 +2,6 @@ use std::collections::HashMap;
 use std::sync::Arc;
 
 use object_store::azure::{AzureConfigKey, MicrosoftAzure, MicrosoftAzureBuilder};
-use object_store::list::{PaginatedListOptions, PaginatedListResult, PaginatedListStore};
 use object_store::ObjectStoreScheme;
 use pyo3::prelude::*;
 use pyo3::pybacked::PyBackedStr;
@@ -487,15 +486,4 @@ fn parse_url(config: Option<PyAzureConfig>, parsed: &Url) -> object_store::Resul
     }
 
     Ok(config)
-}
-
-#[async_trait::async_trait]
-impl PaginatedListStore for PyAzureStore {
-    async fn list_paginated(
-        &self,
-        prefix: Option<&str>,
-        opts: PaginatedListOptions,
-    ) -> object_store::Result<PaginatedListResult> {
-        self.store.list_paginated(prefix, opts).await
-    }
 }
